add a callback feature and change to a clearer token based system
when requesting a wh we can put tokens on it, and whenever we acquire (or
test) it we can withdraw these. This helps to keep track if a wh is still
needed in the queue or can be released.

This token system allows us to implement a callback feature. There a
callback thread (with a user defined function) waits until the wh is
acquired and than launches this one desired action.