add a data field to orwl_wq
this will be used as a sort of shared variable
access to this is through orwl_wh (and derivatives) and
after the handle has been acquired

The idea is to promote the actual value from the lock queue to the one
that acquires. A modified value is only returned on release if the
request had been for reading.