bugfix in the TMIN / TMAX macros
actually `>>' for signed types can be arithmetic or logic, so the
resulting value of TMIN was not well defined; in fact for GCC it was the
wrong one.

Here we give a new implementation of the two that only uses `<<' which is
always defined to be the logic shift operator.