add quasi-full support for the icc
Now we have 4 compilers for which p99 and orwl are tested

- gcc
- clang
- opencc
- icc

Among these opencc and icc have still some rough edges, they have fallen
quite behind their pretended gcc support. In particular icc pretends to
be compatible with whatever gcc version is found on the system, but this
is a plain lie.