make the block macros robust for arguments that contain commas
If a final argument is supposed to receive commas, we just use
``...'' instead of a named argument.

If another argument may receive commas we have to protect it with
P00_ROBUST. By that the commas in the argument are first skipped when the
preprocessor seeks the individual arguments because they are protected by
``( )''. Then, when the individual arguments are expanded the syntactic
sugar of ``P00_ROBUST( )'' simply disappears and the argument with its
commas is passed through.