replace the mutex/condition in orwl_wh by lockfree operations
The thread that puts an element in front of the queue now marks the
element as having acquired the lock. The reading of the flag is atomic
and threads that must ensure that the lock is acquired can just block on
this flag.

This now allows to have cleaner invariants:
- the order in the orwl_wq is only changed when the mutex of the wq is
  hold
- Checking, verifying and reference counting can be done atomically
- A lock is acquired iff the flag is set