some cleanup of orwl_wq and orwl_wh
in particular, ensure that we only use realloc for the data
(re)-allocation. This has a bit of disadvantage that it is allowed to
return something else than 0 when called with a size of 0. On the other
hand this will allow us to replace realloc by just one other allocation
routine if we want to allocate special types of memory.