introduce scoped variables for macro block expressions à la gcc
macro expressions with ({ .... }) are a powerful tool if used together
with the __typeof__ extension. But there is a danger that in something
like

__typeof__(BLA) p00_bla = (BLA);

the variable name p00_bla is already used on the right hand side, hidden
in the evaluation of the macro parameter BLA.

We introduce a macro that first evaluates all macro arguments in dummy
variables and then creates variables with names given by the programmer
for all these values. By that we can freely dispose of names macro for
macro parameters.