implement _Atomic also for pointer types
as a side effect have _Atomic operations also lock-free for all types
that have the same size as some unsigned integer type, with the notable
exception of floating point types. For those we have to use an
atomic_flag to protect them if they are used with atomic arithmetic
operations.

also classify the lock-free types much more systematically